Sanitize phx-change attribute in formsForRecovery
Fixes #2252

When using the JS module to set phx-change, the call to
formsForRecovery on frontend will fail due to not passing the attribute
value correctly into a css selector.

Escaping is necessary for the css selector to become valid.

CSS.escape would be a more foolproof approach but jest does not support
it in node env in which it runs. Not sure if it supports it in jsdom.

In that case, our best bet would be to stub CSS.escape and spy on it